requirements.txt
et-xmlfile==1.1.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
pyenchant==3.2.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
